Add product search feature using Elasticsearch

+++ database/migrations/2024_08_06_143934_create_carts_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('user_id');
-            $table->float('price');
+            $table->unsignedInteger('price');
             $table->integer('quantity');
-            $table->float('amount');
+            $table->unsignedInteger('amount');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('CASCADE');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE');
             $table->timestamps();
