+++ database/migrations/2024_10_21_095332_create_order_details_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->string('order_number');
             $table->string('slug');
             $table->string('title');
-            $table->float('price');
+            $table->unsignedInteger('price');
             $table->integer('quantity');
-            $table->float('amount');
+            $table->unsignedInteger('amount');
             $table->foreign('order_number')->references('order_number')->on('orders')->onDelete('CASCADE');
             $table->timestamps();
         });
